requirements.txt
dj-database-url==0.4.1
djangorestframework==3.4.0
Django==1.9.7
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2
django-widget-tweaks==1.4.1
django-notifications-hq==1.2
